+++ src/main/java/onlyoffice/integration/ui/EditMenuContext.java
@@ -187,7 +187,7 @@ extends BaseDLViewFileVersionDisplayContext {
 //      MutableRenderParameters params = portletURL.getRenderParameters();
 //      params.setValue("fileId", Long.toString(fileVersion.getFileVersionId()));
 
-        portletURL.setParameter("fileId", String.valueOf(Long.toString(fileVersion.getFileVersionId())));
+        portletURL.setParameter("fileId", Long.toString(fileVersion.getFileEntryId()));
 
         try {
             portletURL.setWindowState(LiferayWindowState.EXCLUSIVE);
@@ -208,7 +208,7 @@ extends BaseDLViewFileVersionDisplayContext {
 //      MutableRenderParameters params = portletURL.getRenderParameters();
 //      params.setValue("fileId", Long.toString(fileVersion.getFileVersionId()));
 
-        portletURL.setParameter("fileId", String.valueOf(Long.toString(fileVersion.getFileVersionId())));
+        portletURL.setParameter("fileId", String.valueOf(fileVersion.getFileEntryId()));
 
         try {
             portletURL.setWindowState(LiferayWindowState.POP_UP);